# Various small optimizations and refactorings