--- a/raphael-solver/src/quality_upper_bound_solver/tests.rs
+++ b/raphael-solver/src/quality_upper_bound_solver/tests.rs
@@ -46,7 +46,7 @@ fn check_consistency(solver_settings: SolverSettings) {
         let state = random_state(&solver_settings, &mut rng);
         let state_upper_bound = solver.quality_upper_bound(state).unwrap();
         for action in FULL_SEARCH_ACTIONS {
-            let child_upper_bound = match use_action_combo(&solver_settings, state, *action) {
+            let child_upper_bound = match use_action_combo(&solver_settings, state, action) {
                 Ok(child) => match child.is_final(&solver_settings.simulator_settings) {
                     false => solver.quality_upper_bound(child).unwrap(),
                     true if child.progress >= u32::from(solver_settings.max_progress()) => {