Measurements with `perf stat` for `macro_solver_example`:

Before:
```
         24.413,53 msec task-clock                       #    1,940 CPUs utilized             
            23.030      context-switches                 #  943,330 /sec                      
             1.847      cpu-migrations                   #   75,655 /sec                      
           222.075      page-faults                      #    9,096 K/sec                     
    92.175.601.000      cycles                           #    3,776 GHz                       
    20.390.910.968      stalled-cycles-frontend          #   22,12% frontend cycles idle      
    91.374.651.268      instructions                     #    0,99  insn per cycle            
                                                  #    0,22  stalled cycles per insn   
    16.338.786.710      branches                         #  669,251 M/sec                     
       816.021.957      branch-misses                    #    4,99% of all branches

      12,586939718 seconds time elapsed
```

After:
```
         23.770,81 msec task-clock                       #    1,985 CPUs utilized             
            24.460      context-switches                 #    1,029 K/sec                     
             1.872      cpu-migrations                   #   78,752 /sec                      
           235.189      page-faults                      #    9,894 K/sec                     
    89.703.146.495      cycles                           #    3,774 GHz                       
    16.861.604.587      stalled-cycles-frontend          #   18,80% frontend cycles idle      
    93.843.832.337      instructions                     #    1,05  insn per cycle            
                                                  #    0,18  stalled cycles per insn   
    15.809.251.679      branches                         #  665,070 M/sec                     
       665.274.602      branch-misses                    #    4,21% of all branches           

      11,974739415 seconds time elapsed
```